2195: Sample cards helper refactoring
